feat: rename page components for consistency and improve layout structure
