+++ app/dashboard/layout.tsx
@@ -14,7 +14,9 @@ export default async function DashboardRootLayout({
       <AppSideBar />
       <SidebarInset>
         <Header />
-        <main>{children}</main>
+        <main className="flex-1 px-4 sm:px-6 lg:px-8 py-6">
+          {children}
+        </main>{" "}
       </SidebarInset>
     </SidebarProvider>
     // </ProtectedRoute>
