+++ components/dashboard/KpiCards.tsx
@@ -17,7 +17,6 @@ const statsCards = [
   },
 ];
 async function KpiCards() {
-  await new Promise((resolve) => setTimeout(resolve, 1000));
   return (
     <div className="grid grid-cols-1 gap-5 sm:grid-cols-2 lg:grid-cols-4 ">
       {statsCards.map((card) => (
